+++ ark/schema/config.ts
@@ -1,4 +1,10 @@
-import type { array, mutable, requireKeys, show } from "@ark/util"
+import {
+	$ark,
+	type array,
+	type mutable,
+	type requireKeys,
+	type show
+} from "@ark/util"
 import type { Ark } from "./keywords/keywords.js"
 import type { IntrinsicKeywords, RawRootScope } from "./scope.js"
 import type {
