+++ ark/schema/constraint.ts
@@ -1,4 +1,5 @@
 import {
+	$ark,
 	append,
 	appendUnique,
 	capitalize,
