+++ ark/schema/generic.ts
@@ -1,4 +1,5 @@
 import {
+	$ark,
 	cached,
 	Callable,
 	flatMorph,
