+++ ark/schema/index.ts
@@ -32,11 +32,13 @@ export * from "./roots/root.js"
 export * from "./roots/union.js"
 export * from "./roots/unit.js"
 export * from "./scope.js"
+export * from "./shared/compile.js"
 export * from "./shared/declare.js"
 export * from "./shared/disjoint.js"
 export * from "./shared/errors.js"
 export * from "./shared/implement.js"
 export * from "./shared/intersections.js"
+export * from "./shared/registry.js"
 export * from "./shared/utils.js"
 export * from "./structure/indexed.js"
 export * from "./structure/optional.js"
