+++ ark/schema/roots/proto.ts
@@ -1,4 +1,5 @@
 import {
+	$ark,
 	builtinConstructors,
 	constructorExtends,
 	getExactBuiltinConstructorName,
