+++ ark/schema/roots/root.ts
@@ -13,6 +13,7 @@ import type {
 	UnknownRangeSchema
 } from "@ark/schema"
 import {
+	$ark,
 	cached,
 	includes,
 	omit,
