+++ ark/schema/structure/shared.ts
@@ -1,8 +1,11 @@
-import { registeredReference } from "@ark/util"
+import {
+	registeredReference,
+	type RegisteredReference
+} from "../shared/registry.js"
 
 export const arrayIndexSource = `^(?:0|[1-9]\\d*)$`
 
 export const arrayIndexMatcher = new RegExp(arrayIndexSource)
 
-export const arrayIndexMatcherReference: `$ark.${string}` =
+export const arrayIndexMatcherReference: RegisteredReference =
 	registeredReference(arrayIndexMatcher)
