+++ ark/util/registry.ts
@@ -1,33 +1,30 @@
-import { domainOf, hasDomain } from "./domain.js"
-import { throwError, throwInternalError } from "./errors.js"
+import { domainOf } from "./domain.js"
+import { throwInternalError } from "./errors.js"
 import { objectKindOf } from "./objectKinds.js"
-import { serializePrimitive, type SerializablePrimitive } from "./primitive.js"
+import packageJson from "./package.json"
 import type { PartialRecord } from "./records.js"
 
-declare global {
-	export const $ark: ArkEnv.registry
+export type InitialRegistryContents = {
+	version: string
+	filename: string
+}
 
+declare global {
 	export interface ArkEnv {
 		registry(): {}
 	}
 
 	export namespace ArkEnv {
 		export type registry = PartialRecord<string, object | symbol> &
+			InitialRegistryContents &
 			ReturnType<ArkEnv["registry"]>
 	}
 }
 
-if ("$ark" in globalThis) {
-	throwError(
-		`Tried to initialize an $ark registry but one already existed.
-This probably means you are either depending on multiple versions of an arktype package,
-or importing the same package from both ESM and CJS.
-Review package.json versions across your repo to ensure consistency.`
-	)
-}
-
-export const registry: Record<string, unknown> = {}
-;(globalThis as any).$ark = registry
+export const $ark: ArkEnv.registry = {
+	version: packageJson.version,
+	filename: import.meta.filename
+} satisfies InitialRegistryContents as never
 
 const namesByResolution = new WeakMap<object | symbol, string>()
 const nameCounts: Record<string, number | undefined> = {}
@@ -40,26 +37,13 @@ export const register = (value: object | symbol): string => {
 	if (nameCounts[name]) name = `${name}${nameCounts[name]!++}`
 	else nameCounts[name] = 1
 
-	registry[name] = value
+	$ark[name] = value
 	namesByResolution.set(value, name)
 	return name
 }
 
-export const reference = (name: string): RegisteredReference => `$ark.${name}`
-
-export const registeredReference = (
-	value: object | symbol
-): RegisteredReference => reference(register(value))
-
-export type RegisteredReference<to extends string = string> = `$ark.${to}`
-
 export const isDotAccessible = (keyName: string): boolean =>
 	/^[a-zA-Z_$][a-zA-Z_$0-9]*$/.test(keyName)
-
-export const compileSerializedValue = (value: unknown): string =>
-	hasDomain(value, "object") || typeof value === "symbol" ?
-		registeredReference(value)
-	:	serializePrimitive(value as SerializablePrimitive)
 
 const baseNameFor = (value: object | symbol) => {
 	switch (typeof value) {
